refactor: Widen property index from `u8` to `u16` across internal API
The extended property services (`A_PropertyExtDescription_Read`) use
12-bit property indices (0-4095), but the internal API used `u8`,
clamping indices >255 to "not found".

Widen `prop_idx` to `u16` in:
- `PropertyDescriptionResponse::prop_idx` and `from_descriptor()`
- `PropertyServiceHandler::property_description_read()`
- `PropertyLookup::ByIndex`
- `InterfaceObject::property_description()`
- All augment implementations (Security, IP, TP1)
- `base_property_count()` return type

The non-ext wire format (`A_PropertyDescription_Read`) remains u8 at
the boundary — widened to u16 at the call site in `application.rs`.
The ext response encoding now correctly writes 12-bit prop_idx across
bytes 7-8 instead of only byte 8.

+++ crates/zweidraehte-device/src/bcus/system_b/extensions/ip/augment.rs
@@ -290,8 +290,8 @@ impl<P: IpPlatform, const N: usize, const CAPS: u16> IpAugment<'_, P, N, CAPS> {
     }
 
     /// Look up a property descriptor by augment-local 0-based index.
-    fn ip_descriptor_by_index(&self, idx: u8) -> Option<PropertyDescriptor> {
-        let base_len = BASE_PROPS.len() as u8;
+    fn ip_descriptor_by_index(&self, idx: u16) -> Option<PropertyDescriptor> {
+        let base_len = BASE_PROPS.len() as u16;
         if idx < base_len {
             return Some(BASE_PROPS[idx as usize]);
         }
